+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+ * Turn `Mat.createFrom()` into a static factory method, and make `Mat.copyFrom()` call `Mat.create()` as appropriate ([issue #1](https://github.com/bytedeco/javacpp-presets/issues/1))
  * Add missing `native_camera` modules of `opencv_highgui` for Android
  * Fix functions from `opencv_stitching` not accepting a `MatVector` as apparently intended by the API (issue javacv:466)